backend/pyproject.toml
[project]
name = "tmp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi",
    "uvicorn",
    "pydantic-ai",
    "google-generativeai",
    "python-dotenv",
]
